Fix more warnings.
We're down to fewer than 100 left now. Yay!

+++ sys/src/cmd/iozone/fileop.c
@@ -1121,7 +1121,7 @@ be2vlong(int64_t *to, char *f)
 	t = (char*)to;
 	o = (char*)&order;
 	for(i = 0; i < 8; i++)
-	t[o[i]] = f[i];
+		t[(int)o[i]] = f[i];
 }
 
 void
